Remove kdl_parser_py.
This is being moved to a new repository.  While we are
doing this, move kdl_parser to the top-level, since it
is the only package in here now.

No functional changes are being made to the package.
