+++ README.md
@@ -1,11 +1,6 @@
 # KDL Parser
 
-This contains packages for converting from URDF to a representation in KDL.
-See the ROS wiki for API documentation and tutorials.
-
-* [`kdl_parser`](http://wiki.ros.org/kdl_parser)
-* [`kdl_parser_py`](http://wiki.ros.org/kdl_parser_py)
+This contains a C++ package for converting from URDF to a representation in KDL.
 
 This was originally part of the [`ros/robot_model`](https://github.com/ros/robot_model) repository.
 It has been moved to this repo as described by [`ros/robot_model#195`](https://github.com/ros/robot_model/issues/195)
-
