--- kdl_parser_py/CHANGELOG.rst
@@ -1,265 +1,0 @@
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-Changelog for package kdl_parser_py
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-
-2.1.0 (2018-06-26)
-------------------
-* point to the source and bugtracker used in ros2 (`#3 <https://github.com/ros2/kdl_parser/issues/3>`_)
-* Contributors: Mikael Arguedas
-
-1.12.10 (2017-05-17)
---------------------
-
-1.12.9 (2017-04-26)
--------------------
-* Unset origin xyz or rpy is now properly interpreted as zero vector (`#187 <https://github.com/ros/robot_model/issues/187>`_)
-* Contributors: eugene-katsevman
-
-1.12.8 (2017-03-27)
--------------------
-* Switch a couple more packages over to Chris and Shane.
-* Contributors: Chris Lalancette
-
-1.12.7 (2017-01-26)
--------------------
-
-1.12.6 (2017-01-04)
--------------------
-
-1.12.5 (2016-10-27)
--------------------
-
-1.12.4 (2016-08-23)
--------------------
-
-1.12.3 (2016-06-10)
--------------------
-
-1.12.2 (2016-04-12)
--------------------
-
-1.12.1 (2016-04-10)
--------------------
-* Remove cmake_modules dependency
-* Contributors: Jackie Kay
-
-1.12.0 (2016-04-04)
--------------------
-
-1.11.10 (2016-02-23)
---------------------
-* another patch release
-* Remove cmake_modules dependency
-* Contributors: Jackie Kay
-
-1.11.9 (2016-02-22)
--------------------
-* Changelog
-* kdl_parser_py: run_depend on urdfdom_py
-* Remove dependency in urdf_parser_py
-* kdl_parser: Adding python kdl parser
-* Contributors: Jackie Kay, Jonathan Bohren, Steven Peters
-
-1.11.8 (2015-09-11)
--------------------
-
-1.11.7 (2015-04-22)
--------------------
-
-1.11.6 (2014-11-30 11:17)
--------------------------
-
-1.11.5 (2014-07-24)
--------------------
-
-1.11.4 (2014-07-07)
--------------------
-
-1.11.3 (2014-06-24)
--------------------
-
-1.11.2 (2014-03-22)
--------------------
-
-1.11.1 (2014-03-20)
--------------------
-
-1.11.0 (2014-02-21)
--------------------
-
-1.10.21 (2014-11-30 11:15)
---------------------------
-
-1.10.20 (2014-08-01)
---------------------
-
-1.10.19 (2014-02-15)
---------------------
-
-1.10.18 (2013-12-04)
---------------------
-
-1.10.17 (2013-11-22)
---------------------
-
-1.10.16 (2013-11-18)
---------------------
-
-1.10.15 (2013-08-17)
---------------------
-
-1.10.14 (2013-07-26)
---------------------
-
-1.10.13 (2013-07-17)
---------------------
-
-1.10.12 (2013-07-02 08:41)
---------------------------
-
-1.10.11 (2013-05-21)
---------------------
-
-1.10.10 (2013-05-20)
---------------------
-
-1.10.9 (2013-04-21)
--------------------
-
-1.10.8 (2013-04-18 12:24)
--------------------------
-
-1.10.7 (2013-04-18 10:19)
--------------------------
-
-1.10.6 (2013-04-18 01:56)
--------------------------
-
-1.10.5 (2013-04-18 01:46)
--------------------------
-
-1.10.4 (2013-04-12)
--------------------
-
-1.10.3 (2013-03-13 22:22)
--------------------------
-
-1.10.2 (2013-03-13 17:34)
--------------------------
-
-1.10.1 (2013-03-13 17:15)
--------------------------
-
-1.10.0 (2013-03-11 19:48)
--------------------------
-
-1.9.36 (2013-07-02 08:38)
--------------------------
-
-1.9.35 (2013-04-29)
--------------------
-
-1.9.34 (2013-04-18 18:17)
--------------------------
-
-1.9.33 (2013-03-11 19:49)
--------------------------
-
-1.9.32 (2012-12-22)
--------------------
-
-1.9.31 (2012-12-18)
--------------------
-
-1.9.30 (2012-12-14)
--------------------
-
-1.9.29 (2012-12-05)
--------------------
-
-1.9.28 (2012-11-07)
--------------------
-
-1.9.27 (2012-11-06)
--------------------
-
-1.9.26 (2012-11-05)
--------------------
-
-1.9.25 (2012-10-29)
--------------------
-
-1.9.24 (2012-10-25)
--------------------
-
-1.9.23 (2012-10-14 15:26)
--------------------------
-
-1.9.22 (2012-10-14 13:13)
--------------------------
-
-1.9.21 (2012-10-14 12:25)
--------------------------
-
-1.9.20 (2012-10-14 02:13)
--------------------------
-
-1.9.19 (2012-10-13)
--------------------
-
-1.9.18 (2012-10-07)
--------------------
-
-1.9.17 (2012-10-06 21:27)
--------------------------
-
-1.9.16 (2012-10-06 21:22)
--------------------------
-
-1.9.15 (2012-10-06 20:47)
--------------------------
-
-1.9.14 (2012-10-06 19:20)
--------------------------
-
-1.9.13 (2012-09-16 16:51)
--------------------------
-
-1.9.12 (2012-09-16 02:25)
--------------------------
-
-1.9.11 (2012-09-15 13:45)
--------------------------
-
-1.9.10 (2012-09-15 12:27)
--------------------------
-
-1.9.9 (2012-09-12 14:38)
-------------------------
-
-1.9.8 (2012-09-12 14:27)
-------------------------
-
-1.9.7 (2012-09-11)
-------------------
-
-1.9.6 (2012-09-07)
-------------------
-
-1.9.5 (2012-09-06)
-------------------
-
-1.9.4 (2012-09-04)
-------------------
-
-1.9.3 (2012-09-03)
-------------------
-
-1.9.2 (2012-08-14 20:34)
-------------------------
-
-1.9.1 (2012-08-14 20:33)
-------------------------
-
-1.9.0 (2012-08-02)
-------------------
