--- kdl_parser_py/test/test_kdl_parser.launch
@@ -1,3 +1,0 @@
-<launch>
-  <test test-name="test_kdl_parser" pkg="kdl_parser_py" type="test_kdl_parser.py" args="$(find kdl_parser_py)/test/test.urdf"/>
-</launch>
